initial changes for ntddk port
These changes are still in the prototype stage.  I will need to
better handle include_next, as well as a number of things that
come from c99 that are not present in the wdk.  This gets
most of the headers compiling though.  <atomic> is the only
standalone header that doesn't compile, though <cstdalign> is
missing entirely.

I need to provide an <atomic> implementation, do a bunch of
cleanups, then start adding in extra headers that would be nice
to have.